Misc fixes

+++ imgui-application/arvt_helpers.cpp
@@ -153,8 +153,13 @@ int copyFileToCStr(const char* path, char* dest, int buf_size) {
 
 int copy_file(const char* path, const char* newPath) {
 	//TODO: this can throw
-	std::filesystem::copy_file(path, newPath, std::filesystem::copy_options::overwrite_existing);
-	return 0;
+	if (( std::filesystem::exists(path)    && std::filesystem::is_regular_file(path)) &&
+	    (!std::filesystem::exists(newPath) || std::filesystem::is_regular_file(newPath))) {
+		std::filesystem::copy_file(path, newPath, std::filesystem::copy_options::overwrite_existing);
+		return 0;
+	} else {
+		return 1;
+	}
 }
 
 //note: this also checks for existence using <filesystem>
